refactoring sections chosen
